change the code structure
