+++ admob-latest/src/main/AndroidManifest.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.rakuten.tech.mobile.admob"
-    android:versionCode="1"
-    android:versionName="0.0">
+    package="com.rakuten.tech.mobile.admob">
 
 </manifest>
