--- admob-latest/src/main/java/com/rakuten/tech/mobile/admob/AdStatus.kt
@@ -1,11 +1,0 @@
-package com.rakuten.tech.mobile.admob
-
-/**
- * Ad Status send from the callbacks.
- */
-enum class AdStatus {
-    LOADED,
-    SHOWED,
-    CLOSED,
-    FAILED,
-}
