+++ admob-latest/src/test/AndroidManifest.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.rakuten.tech.mobile.admob.legacy.platform"
-    android:versionCode="1"
-    android:versionName="0.0">
+    package="com.rakuten.tech.mobile.admob.legacy.platform">
     <application>
         <activity android:name="com.rakuten.tech.mobile.admob.TestActivity"/>
     </application>
